+++ packages/clerk-js/src/ui/components/PaymentSources/AddPaymentSource.tsx
@@ -15,7 +15,7 @@ import { FormButtons } from '@/ui/elements/FormButtons';
 import { FormContainer } from '@/ui/elements/FormContainer';
 
 import { clerkUnsupportedEnvironmentWarning } from '../../../core/errors';
-import { useEnvironment, useSubscriberTypeContext } from '../../contexts';
+import { useEnvironment, useSubscriberTypeContext, useSubscriberTypeLocalizationRoot } from '../../contexts';
 import { descriptors, Flex, localizationKeys, Spinner, useAppearance, useLocalizations } from '../../customizables';
 import type { LocalizationKey } from '../../localization';
 import { handleError, normalizeColorString } from '../../utils';
@@ -236,6 +236,7 @@ const AddPaymentSourceForm = ({ children }: PropsWithChildren) => {
   const elements = useElements();
   const { displayConfig } = useEnvironment();
   const { t } = useLocalizations();
+  const localizationRoot = useSubscriberTypeLocalizationRoot();
 
   const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -260,10 +261,8 @@ const AddPaymentSourceForm = ({ children }: PropsWithChildren) => {
     try {
       await onSuccess({ stripeSetupIntent: setupIntent });
     } catch (error) {
-      console.log('catch', error);
       void handleError(error, [], card.setError);
     } finally {
-      console.log('finally');
       card.setIdle();
       initializePaymentSource(); // resets the payment intent
     }
@@ -290,8 +289,7 @@ const AddPaymentSourceForm = ({ children }: PropsWithChildren) => {
               type: 'tabs',
               defaultCollapsed: false,
             },
-            // TODO(@COMMERCE): Should this be fetched from the fapi?
-            paymentMethodOrder: paymentMethodOrder || ['card'],
+            paymentMethodOrder,
             applePay: checkout
               ? {
                   recurringPaymentRequest: {
@@ -311,7 +309,8 @@ const AddPaymentSourceForm = ({ children }: PropsWithChildren) => {
         <FormButtons
           isDisabled={!isPaymentElementReady}
           submitLabel={
-            submitLabel ?? localizationKeys('userProfile.billingPage.paymentSourcesSection.formButtonPrimary__add')
+            submitLabel ??
+            localizationKeys(`${localizationRoot}.billingPage.paymentSourcesSection.formButtonPrimary__add`)
           }
           onReset={cancelAction}
           hideReset={!cancelAction}