+++ packages/localizations/src/de-DE.ts
@@ -14,6 +14,28 @@ import type { LocalizationResource } from '@clerk/types';
 
 export const deDE: LocalizationResource = {
   locale: 'de-DE',
+  apiKeys: {
+    action__add: 'Neuen API-Key hinzufügen',
+    action__search: 'Suche',
+    dates: {
+      lastUsed__days: undefined,
+      lastUsed__hours: undefined,
+      lastUsed__minutes: undefined,
+      lastUsed__months: undefined,
+      lastUsed__seconds: undefined,
+      lastUsed__years: undefined,
+    },
+    detailsTitle__emptyRow: 'Keine API-Keys gefunden',
+    formButtonPrimary__add: 'API-Key erstellen',
+    formHint: 'Geben Sie einen Namen an, um einen API-Key zu erstellen. Sie können ihn jederzeit widerrufen.',
+    formTitle: 'Neuen API-Key hinzufügen',
+    menuAction__revoke: 'API-Key widerrufen',
+    revokeConfirmation: {
+      formButtonPrimary__revoke: 'API-Key widerrufen',
+      formHint: 'Sind Sie sicher, dass Sie diesen API-Key löschen wollen?',
+      formTitle: 'API-Key "{{apiKeyName}}" widerrufen?',
+    },
+  },
   backButton: 'Zurück',
   badge__activePlan: 'Aktiv',
   badge__canceledEndsAt: "Storniert • Endet am {{ date | shortDate('de-DE') }}",
@@ -65,6 +87,7 @@ export const deDE: LocalizationResource = {
       title__paymentSuccessful: 'Zahlung erfolgreich!',
       title__subscriptionSuccessful: 'Geschafft!',
     },
+    credit: undefined,
     creditRemainder: 'Verbleibendes Guthaben für den restlichen Abrechnungszeitraum.',
     defaultFreePlanActive: 'Sie nutzen aktuell den kostenlosen Plan.',
     free: 'Kostenlos',
@@ -138,6 +161,9 @@ export const deDE: LocalizationResource = {
   formFieldHintText__optional: 'Optional',
   formFieldHintText__slug:
     'Der Slug ist eine für Menschen lesbare ID. Sie muss einzigartig sein und wird oft in URLs verwendet.',
+  formFieldInputPlaceholder__apiKeyDescription: 'Geben Sie eine Beschreibung an',
+  formFieldInputPlaceholder__apiKeyExpirationDate: 'Geben Sie ein Ablaufdatum an',
+  formFieldInputPlaceholder__apiKeyName: 'Geben Sie einen Namen an',
   formFieldInputPlaceholder__backupCode: 'Sicherheitscode eingeben',
   formFieldInputPlaceholder__confirmDeletionUserAccount: 'Konto löschen',
   formFieldInputPlaceholder__emailAddress: 'E-Mail-Adresse eingeben',
@@ -152,6 +178,10 @@ export const deDE: LocalizationResource = {
   formFieldInputPlaceholder__password: 'Passwort eingeben',
   formFieldInputPlaceholder__phoneNumber: 'Telefonnummer eingeben',
   formFieldInputPlaceholder__username: 'Benutzername eingeben',
+  formFieldLabel__apiKeyDescription: 'Beschreibung',
+  formFieldLabel__apiKeyExpiration: 'Ablaufdatum',
+  formFieldLabel__apiKeyExpirationDate: 'Datum auswählen',
+  formFieldLabel__apiKeyName: 'Name',
   formFieldLabel__automaticInvitations: 'Aktivieren Sie automatische Einladungen für diese Domain',
   formFieldLabel__backupCode: 'Sicherungscode',
   formFieldLabel__confirmDeletion: 'Bestätigung',
@@ -197,6 +227,9 @@ export const deDE: LocalizationResource = {
     titleWithoutPersonal: 'Organisation auswählen',
   },
   organizationProfile: {
+    apiKeysPage: {
+      title: 'API-Keys',
+    },
     badge__automaticInvitation: 'Automatische Einladungen',
     badge__automaticSuggestion: 'Automatische Vorschläge',
     badge__manualInvitation: 'Keine automatische Aufnahme',
@@ -221,6 +254,7 @@ export const deDE: LocalizationResource = {
         title: 'Zahlungsmethoden',
       },
       start: {
+        headerTitle__payments: undefined,
         headerTitle__plans: 'Pläne',
         headerTitle__statements: 'Abrechnungen',
         headerTitle__subscriptions: 'Abonnements',
@@ -295,6 +329,7 @@ export const deDE: LocalizationResource = {
       },
     },
     navbar: {
+      apiKeys: 'API-Keys',
       billing: 'Abrechnung',
       description: 'Verwalten Sie ihre Organisation.',
       general: 'Allgemein',
@@ -825,6 +860,9 @@ export const deDE: LocalizationResource = {
     action__signOutAll: 'Melden Sie sich von allen Konten ab',
   },
   userProfile: {
+    apiKeysPage: {
+      title: 'API-Keys',
+    },
     backupCodePage: {
       actionLabel__copied: 'Kopiert!',
       actionLabel__copy: 'Kopiere alle',
@@ -861,6 +899,7 @@ export const deDE: LocalizationResource = {
         title: 'Zahlungsmethoden',
       },
       start: {
+        headerTitle__payments: undefined,
         headerTitle__plans: 'Pläne',
         headerTitle__statements: 'Abrechnungen',
         headerTitle__subscriptions: 'Abonnements',
@@ -987,6 +1026,7 @@ export const deDE: LocalizationResource = {
     mobileButton__menu: 'Menü',
     navbar: {
       account: 'Profil',
+      apiKeys: 'API-Keys',
       billing: 'Abrechnung',
       description: 'Verwalten Sie Ihre Kontoinformationen.',
       security: 'Sicherheit',