+++ packages/types/src/localization.ts
@@ -54,6 +54,10 @@ type _LocalizationResource = {
   formFieldLabel__confirmDeletion: LocalizationValue;
   formFieldLabel__role: LocalizationValue;
   formFieldLabel__passkeyName: LocalizationValue;
+  formFieldLabel__apiKeyName: LocalizationValue;
+  formFieldLabel__apiKeyDescription: LocalizationValue;
+  formFieldLabel__apiKeyExpiration: LocalizationValue;
+  formFieldLabel__apiKeyExpirationDate: LocalizationValue;
   formFieldInputPlaceholder__emailAddress: LocalizationValue;
   formFieldInputPlaceholder__emailAddresses: LocalizationValue;
   formFieldInputPlaceholder__phoneNumber: LocalizationValue;
@@ -68,6 +72,9 @@ type _LocalizationResource = {
   formFieldInputPlaceholder__organizationDomain: LocalizationValue;
   formFieldInputPlaceholder__organizationDomainEmailAddress: LocalizationValue;
   formFieldInputPlaceholder__confirmDeletionUserAccount: LocalizationValue;
+  formFieldInputPlaceholder__apiKeyName: LocalizationValue;
+  formFieldInputPlaceholder__apiKeyDescription: LocalizationValue;
+  formFieldInputPlaceholder__apiKeyExpirationDate: LocalizationValue;
   formFieldError__notMatchingPasswords: LocalizationValue;
   formFieldError__matchingPasswords: LocalizationValue;
   formFieldError__verificationLinkExpired: LocalizationValue;
@@ -131,6 +138,7 @@ type _LocalizationResource = {
     subtotal: LocalizationValue;
     credit: LocalizationValue;
     creditRemainder: LocalizationValue;
+    totalDue: LocalizationValue;
     totalDueToday: LocalizationValue;
     pastDue: LocalizationValue;
     paymentMethods: LocalizationValue;
@@ -498,6 +506,7 @@ type _LocalizationResource = {
       account: LocalizationValue;
       security: LocalizationValue;
       billing: LocalizationValue;
+      apiKeys: LocalizationValue;
     };
     start: {
       headerTitle__account: LocalizationValue;
@@ -638,6 +647,10 @@ type _LocalizationResource = {
         successMessage: LocalizationValue;
       };
     };
+    apiKeysPage: {
+      title: LocalizationValue;
+      detailsTitle__emptyRow: LocalizationValue;
+    };
     passkeyScreen: {
       title__rename: LocalizationValue;
       subtitle__rename: LocalizationValue;
@@ -762,13 +775,34 @@ type _LocalizationResource = {
         headerTitle__subscriptions: LocalizationValue;
         headerTitle__statements: LocalizationValue;
       };
+      statementsSection: {
+        empty: LocalizationValue;
+        itemCaption__paidForPlan: LocalizationValue;
+        itemCaption__proratedCredit: LocalizationValue;
+        itemCaption__subscribedAndPaidForPlan: LocalizationValue;
+        notFound: LocalizationValue;
+        tableHeader__date: LocalizationValue;
+        tableHeader__amount: LocalizationValue;
+        title: LocalizationValue;
+        totalPaid: LocalizationValue;
+      };
       switchPlansSection: {
         title: LocalizationValue;
       };
       subscriptionsListSection: {
+        tableHeader__plan: LocalizationValue;
+        tableHeader__startDate: LocalizationValue;
+        tableHeader__edit: LocalizationValue;
         title: LocalizationValue;
         actionLabel__newSubscription: LocalizationValue;
         actionLabel__switchPlan: LocalizationValue;
+      };
+      paymentHistorySection: {
+        empty: LocalizationValue;
+        notFound: LocalizationValue;
+        tableHeader__date: LocalizationValue;
+        tableHeader__amount: LocalizationValue;
+        tableHeader__status: LocalizationValue;
       };
       paymentSourcesSection: {
         title: LocalizationValue;
@@ -824,6 +858,7 @@ type _LocalizationResource = {
       general: LocalizationValue;
       members: LocalizationValue;
       billing: LocalizationValue;
+      apiKeys: LocalizationValue;
     };
     badge__unverified: LocalizationValue;
     badge__automaticInvitation: LocalizationValue;
@@ -967,13 +1002,34 @@ type _LocalizationResource = {
         headerTitle__subscriptions: LocalizationValue;
         headerTitle__statements: LocalizationValue;
       };
+      statementsSection: {
+        empty: LocalizationValue;
+        itemCaption__paidForPlan: LocalizationValue;
+        itemCaption__proratedCredit: LocalizationValue;
+        itemCaption__subscribedAndPaidForPlan: LocalizationValue;
+        notFound: LocalizationValue;
+        tableHeader__date: LocalizationValue;
+        tableHeader__amount: LocalizationValue;
+        title: LocalizationValue;
+        totalPaid: LocalizationValue;
+      };
       switchPlansSection: {
         title: LocalizationValue;
       };
       subscriptionsListSection: {
+        tableHeader__plan: LocalizationValue;
+        tableHeader__startDate: LocalizationValue;
+        tableHeader__edit: LocalizationValue;
         title: LocalizationValue;
         actionLabel__newSubscription: LocalizationValue;
         actionLabel__switchPlan: LocalizationValue;
+      };
+      paymentHistorySection: {
+        empty: LocalizationValue;
+        notFound: LocalizationValue;
+        tableHeader__date: LocalizationValue;
+        tableHeader__amount: LocalizationValue;
+        tableHeader__status: LocalizationValue;
       };
       paymentSourcesSection: {
         title: LocalizationValue;
@@ -1001,6 +1057,10 @@ type _LocalizationResource = {
       alerts: {
         noPermissionsToManageBilling: LocalizationValue;
       };
+    };
+    apiKeysPage: {
+      title: LocalizationValue;
+      detailsTitle__emptyRow: LocalizationValue;
     };
   };
   createOrganization: {
@@ -1042,6 +1102,28 @@ type _LocalizationResource = {
       title: LocalizationValue;
       subtitle: LocalizationValue;
       message: LocalizationValue;
+    };
+  };
+  apiKeys: {
+    formTitle: LocalizationValue;
+    formHint: LocalizationValue;
+    formButtonPrimary__add: LocalizationValue;
+    menuAction__revoke: LocalizationValue;
+    action__search: LocalizationValue;
+    action__add: LocalizationValue;
+    detailsTitle__emptyRow: LocalizationValue;
+    revokeConfirmation: {
+      formTitle: LocalizationValue;
+      formHint: LocalizationValue;
+      formButtonPrimary__revoke: LocalizationValue;
+    };
+    dates: {
+      lastUsed__seconds: LocalizationValue;
+      lastUsed__minutes: LocalizationValue;
+      lastUsed__hours: LocalizationValue;
+      lastUsed__days: LocalizationValue;
+      lastUsed__months: LocalizationValue;
+      lastUsed__years: LocalizationValue;
     };
   };
 };